+++ src/router/index.js
@@ -27,78 +27,78 @@ const routes = [
   {
     path: '/indexStu',
     name: 'indexStu',
-    component: () => import(/* webpackChunkName: "about" */ '../views/Index_stu.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/students/Index_stu.vue'),
   },
   {
     path: '/stuExamDetail',
     name: 'stuExamDetail',
-    component: () => import(/* webpackChunkName: "about" */ '../views/StuExamDetail.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/students/StuExamDetail.vue'),
   },
   {
     path: '/stuExamGrades',
     name: 'stuExamGrades',
-    component: () => import(/* webpackChunkName: "about" */ '../views/StuExamGrades.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/students/StuExamGrades.vue'),
   },
   {
     path: '/GradesCenter_tch',
     name: 'GradesCenter_tch',
-    component: () => import(/* webpackChunkName: "about" */ '../views/GradesCenter_tch.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/teacher/GradesCenter_tch.vue'),
   },
   {
     path: '/GradesCenter_Stu',
     name: 'GradesCenter_Stu',
-    component: () => import(/* webpackChunkName: "about" */ '../views/GradesCenter_Stu.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/students/GradesCenter_Stu.vue'),
   },
   {
     path: '/StuGradesCenter',
     name: 'StuGradesCenter',
-    component: () => import(/* webpackChunkName: "about" */ '../views/StuGradesCenter.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/teacher/StuGradesCenter.vue'),
   },
   {
     path: '/teaCourseDetail',
     name: 'teaCourseDetail',
-    component: () => import('../views/TeaCourseDetail.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../views/teacher/TeaCourseDetail.vue'),
   },
   {
     path: '/indexTch',
     name: 'indexTch',
-    component: () => import(/* webpackChunkName: "about" */'../views/Index_tch.vue'),
+    component: () => import(/* webpackChunkName: "about" */'../views/teacher/Index_tch.vue'),
 
   },
   {
     path: '/personalStu',
     name: 'personalStu',
-    component: () => import('../components/personal_stu.vue'),
+    component: () => import('../components/student/personal_stu.vue'),
   },
   {
     path: '/personalTch',
     name: 'personalTch',
-    component: () => import('../components/personal_tch.vue'),
+    component: () => import('../components/teacher/personal_tch.vue'),
   },
   {
     path: '/SelectCourse',
     name: 'SelectCourse',
-    component: () => import('../views/SelectCourse'),
+    component: () => import('../views/students/SelectCourse'),
   },
   {
     path: '/AddExam',
     name: 'AddExam',
-    component: () => import('../views/AddExam'),
+    component: () => import('../views/teacher/AddExam'),
   },
   {
     path: '/AddQuestion',
     name: 'AddQuestion',
-    component: () => import('../views/AddQuestion'),
+    component: () => import('../views/teacher/AddQuestion'),
   },
   {
     path: '/StuQuestion',
     name: 'StuStudent',
-    component: () => import('../views/StuQuestion'),
+    component: () => import('../views/students/StuQuestion'),
   },
   {
     path: '/ScoreCenter',
     name: 'ScoreCenter',
-    component: () => import('../views/ScoreCenter'),
+    component: () => import('../views/teacher/ScoreCenter'),
   },
   {
     path: '/ExamInfo',
